requirements.txt
alpaca-py>=0.28.0
python-dotenv>=1.0.0
pytz>=2024.1